pyproject.toml
[tool.poetry]
name = "kt-ds-ai-agent"
version = "0.1.0"
description = "KT DS 입찰 AI Agent 통합 스캐폴드"
authors = ["KT DS AI Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.25.0"
pymupdf = "^1.26.1"
python-dotenv = "^1.1.0"
azure-search-documents = "^11.5.2"
numpy = "1.26.4"
requests = "^2.32.4"
pypdf2 = "^3.0.1"
python-docx = "^1.2.0"
mammoth = "^1.9.1"
streamlit = "^1.45.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
